Slices split out to control persistence
